Crashing FIXED, no need for a revert.
The issue was caused by mis-matched texture file names with the in-editor texture names.  The fix is, perversely, to rename them with a new name that has not been seen by the editor before.  Even after exposing and doing a "fix up" on the object redirector that unreal so helpfully added to my project, the name "blackout_1024.png" is forever tainted, and now it needs a new name.  Amazing.  Unreal, even.